Stockage et calcul des couleurs avec des Double, puis recopie des valeurs dans une structure en Float pour l'affichage.
=> Rend possible le LR progressif

+++ include/outils/couleur.h
@@ -4,7 +4,7 @@
 class Couleur {
     public:
         Couleur();
-        Couleur(float r, float g, float b);
+        Couleur(double r, double g, double b);
         Couleur(const Couleur& pixel);
         ~Couleur();
 
@@ -19,7 +19,7 @@ class Couleur {
         */
         Couleur clamp();
 
-        float r, g, b;
+        double r, g, b;
 };
 
 #endif
